Refactor death message build & event params pass
Split the prefix and message body from the completed built death message.
In the past, the death message component pass as one singe TextComponent.
And now, it is changed to TextComponent[], the first element is prefix component,
and the second the death message body.

By separate passing to event, it will build / append during broadcast stage (In BroadcastXXXDeathListener)
and Discord hook message send stage (In each Discord extension hook). Build twice or more for now,
instead of appending them during death message build.

Build twice: or more
1. In broadcastListener: console message + broadcast message to players
2. In DiscordSRV hook: discord format death message sent to discord
3. In other Discord integration hooks will add in the future

There are also duplicate / similar code, but is not a big deal.